+++ runtime/SmartHauling.Runtime/PluginInfo.cs
@@ -4,5 +4,5 @@ internal static class PluginInfo
 {
     internal const string Guid = "hu.tigyi.goingmedieval.smarthauling.runtime";
     internal const string Name = "Smart Hauling Runtime";
-    internal const string Version = "2.1";
+    internal const string Version = "2.2.0";
 }